Move `MessageSender` to test-harness
