+++ test-harness/tests/poll_api.rs
@@ -1,17 +1,12 @@
 use futures::executor::LocalPool;
-use futures::future::{join, BoxFuture};
+use futures::future::join;
 use futures::prelude::*;
-use futures::stream::FuturesUnordered;
 use futures::task::{Spawn, SpawnExt};
-use futures::{
-    future, stream, AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt, FutureExt, StreamExt,
-};
+use futures::{future, stream, AsyncReadExt, AsyncWriteExt, FutureExt, StreamExt};
 use quickcheck::QuickCheck;
-use std::future::Future;
 use std::iter;
 use std::panic::panic_any;
-use std::pin::{pin, Pin};
-use std::task::{Context, Poll};
+use std::pin::pin;
 use test_harness::*;
 use tokio::net::TcpStream;
 use tokio::runtime::Runtime;
@@ -282,88 +277,4 @@ fn write_deadlock() {
             )
             .unwrap(),
     );
-}
-
-struct MessageSender<T> {
-    connection: Connection<T>,
-    pending_messages: Vec<Msg>,
-    worker_streams: FuturesUnordered<BoxFuture<'static, ()>>,
-    streams_processed: usize,
-    /// Whether to spawn a new task for each stream.
-    spawn_tasks: bool,
-}
-
-impl<T> MessageSender<T> {
-    fn new(connection: Connection<T>, messages: Vec<Msg>, spawn_tasks: bool) -> Self {
-        Self {
-            connection,
-            pending_messages: messages,
-            worker_streams: FuturesUnordered::default(),
-            streams_processed: 0,
-            spawn_tasks,
-        }
-    }
-}
-
-impl<T> Future for MessageSender<T>
-where
-    T: AsyncRead + AsyncWrite + Unpin,
-{
-    type Output = yamux::Result<usize>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
-
-        loop {
-            if this.pending_messages.is_empty() && this.worker_streams.is_empty() {
-                futures::ready!(this.connection.poll_close(cx)?);
-
-                return Poll::Ready(Ok(this.streams_processed));
-            }
-
-            if let Some(message) = this.pending_messages.pop() {
-                match this.connection.poll_new_outbound(cx)? {
-                    Poll::Ready(mut stream) => {
-                        let future = async move {
-                            send_recv_message(&mut stream, message).await.unwrap();
-                            stream.close().await.unwrap();
-                        };
-
-                        let worker_stream_future = if this.spawn_tasks {
-                            async { task::spawn(future).await.unwrap() }.boxed()
-                        } else {
-                            future.boxed()
-                        };
-
-                        this.worker_streams.push(worker_stream_future);
-                        continue;
-                    }
-                    Poll::Pending => {
-                        this.pending_messages.push(message);
-                    }
-                }
-            }
-
-            match this.worker_streams.poll_next_unpin(cx) {
-                Poll::Ready(Some(())) => {
-                    this.streams_processed += 1;
-                    continue;
-                }
-                Poll::Ready(None) | Poll::Pending => {}
-            }
-
-            match this.connection.poll_next_inbound(cx)? {
-                Poll::Ready(Some(stream)) => {
-                    drop(stream);
-                    panic!("Did not expect remote to open a stream");
-                }
-                Poll::Ready(None) => {
-                    panic!("Did not expect remote to close the connection");
-                }
-                Poll::Pending => {}
-            }
-
-            return Poll::Pending;
-        }
-    }
 }
